pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "difference finder"
authors = ["Iaroslav <GunGalla@gmail.com>"]
readme = "README.md"
packages = [{include = "diff_finder"}]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
asciinema = "^2.2.0"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
coverage = "^7.2.1"

[tool.poetry.scripts]
gendiff = "diff_finder.scripts.gendiff:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
